requirements.txt
2captcha-python==1.5.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
distlib==0.3.8
Django==5.1.3
dotenv==0.9.9
ebaysdk @ file:///Users/borisatanasov/Desktop/ebaysdk-python-master
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==2.1.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
piexif==1.1.3
pillow==11.1.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
pyOpenSSL==24.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.1
pytweening==1.2.0
requests==2.32.3
rubicon-objc==0.4.9
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.23.1
service-identity==24.2.0
setuptools==71.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
threadpoolctl==3.5.0
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
whitenoise==6.8.2
wsproto==1.2.0
zope.interface==7.1.1
